reworked pokemon class
I wanted to reduce the amount of input that the user had to give.
ivs or stats can be submitted as a list, with the default parameter as
an empty list. level is also given on initialization. This also makes
more sense so that the getStats anf getHP functions only have to work
with ivs